@@ -26,57 +24,60 @@
     "variant-testing"
   ],
   "dependencies": {
-    "@babel/cli": "^7.15.7",
-    "@babel/core": "^7.15.5",
-    "@babel/plugin-proposal-class-properties": "^7.14.5",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.14.5",
-    "@babel/plugin-proposal-optional-chaining": "^7.14.5",
-    "@babel/plugin-transform-react-jsx": "^7.14.9",
-    "@babel/preset-env": "^7.15.6",
-    "@babel/preset-typescript": "^7.16.0",
+    "@babel/cli": "^7.16.8",
+    "@babel/core": "^7.16.12",
+    "@babel/plugin-proposal-class-properties": "^7.16.7",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.16.7",
+    "@babel/plugin-proposal-optional-chaining": "^7.16.7",
+    "@babel/plugin-transform-react-jsx": "^7.16.7",
+    "@babel/plugin-transform-runtime": "^7.16.10",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-typescript": "^7.16.7",
     "@emotion/babel-plugin-jsx-pragmatic": "^0.1.5",
-    "@rollup/plugin-alias": "^3.1.5",
+    "@rollup/plugin-alias": "^3.1.9",
     "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-commonjs": "^21.0.0",
+    "@rollup/plugin-commonjs": "^21.0.1",
     "@rollup/plugin-image": "^2.1.1",
-    "@rollup/plugin-node-resolve": "^13.0.5",
-    "@rollup/plugin-replace": "^3.0.0",
-    "autoprefixer": "^10.3.6",
-    "babel-plugin-transform-async-to-promises": "^0.8.15",
-    "chalk": "^4.1.2",
-    "chokidar": "^3.5.2",
-    "core-js": "^3.18.1",
-    "dotenv": "^10.0.0",
+    "@rollup/plugin-node-resolve": "^13.1.3",
+    "@rollup/plugin-replace": "^3.0.1",
+    "autoprefixer": "^10.4.2",
+    "babel-plugin-transform-async-to-promises": "^0.8.18",
+    "babel-plugin-transform-import-meta": "^2.1.0",
+    "chalk": "^5.0.0",
+    "chokidar": "^3.5.3",
+    "core-js": "^3.20.3",
+    "dotenv": "^14.3.2",
     "less": "^4.1.2",
     "lodash.merge": "^4.6.2",
     "minimist": "^1.2.5",
-    "node-sass": "^6.0.1",
-    "postcss": "^8.3.8",
-    "postcss-modules": "^4.2.2",
-    "preact": "^10.5.14",
-    "puppeteer-core": "^10.4.0",
-    "rollup": "^2.58.0",
+    "node-sass": "^7.0.1",
+    "postcss": "^8.4.5",
+    "postcss-modules": "^4.3.0",
+    "preact": "^10.6.5",
+    "puppeteer-core": "^13.1.2",
+    "rollup": "^2.66.1",
     "rollup-plugin-iife": "^0.5.0",
     "rollup-plugin-inline-svg": "^2.0.0",
-    "rollup-plugin-postcss": "^4.0.1",
-    "rollup-plugin-styles": "^3.14.1",
+    "rollup-plugin-node-externals": "^3.1.2",
+    "rollup-plugin-postcss": "^4.0.2",
+    "rollup-plugin-styles": "^4.0.0",
     "rollup-plugin-svg-hyperscript": "^0.1.2",
     "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.30.0",
+    "rollup-plugin-typescript2": "^0.31.1",
     "string-hash": "^1.1.3",
     "tslib": "^2.3.1",
-    "typescript": "^4.4.4"
+    "typescript": "^4.5.5"
   },
   "devDependencies": {
-    "expect-puppeteer": "^6.0.0",
+    "babel-jest": "^27.4.6",
+    "expect-puppeteer": "^6.0.2",
     "glob": "^7.2.0",
-    "husky": "^7.0.2",
-    "jest": "^27.2.4",
+    "husky": "^7.0.4",
+    "jest": "^27.4.7",
     "jest-fail-on-console": "^2.1.1",
-    "npm-run-all": "^4.1.5",
-    "puppeteer": "^10.4.0",
+    "puppeteer": "^13.1.2",
     "rimraf": "^3.0.2",
-    "standard-version": "^9.3.1"
+    "standard-version": "^9.3.2"
   },
   "bin": {
     "abdo": "dist/bin/abdo.js"
