@@ -1,13 +1,14 @@
-#!/usr/bin/env node
-import { readdirSync, lstatSync, readFileSync } from 'fs';
+import chalk from 'chalk';
+import chokidar from 'chokidar';
+import minimist from 'minimist';
 import path from 'path';
 import { createFilter } from '@rollup/pluginutils';
+import { readdirSync, lstatSync, readFileSync } from 'fs';
 import buildspec from '../lib/buildspec';
 import { bundler, openPage } from '../lib/bundler';
 import { getBrowser } from '../lib/puppeteer';
-import { cyan, yellow, green, red } from 'chalk';
-import chokidar from 'chokidar';
-import minimist from 'minimist';
+
+const { cyan, yellow, green, red } = chalk;
 
 const cmd = process.argv[2];
 
@@ -61,7 +62,7 @@ switch (cmd) {
  * @param {string} targetPath
  */
 async function buildSingleEntry(targetPath) {
-	const [config] = getMatchingBuildspec(targetPath);
+	const [config] = await getMatchingBuildspec(targetPath);
 
 	if (!config) {
 		console.log(red("Couldn't find buildspec.json for the variant", targetPath));
@@ -115,7 +116,7 @@ async function buildSingleEntry(targetPath) {
  */
 async function buildMultiEntry(targetPath) {
 	const buildOnly = cmd === 'build-all';
-	const buildspecs = getMatchingBuildspec(targetPath);
+	const buildspecs = await getMatchingBuildspec(targetPath);
 	const [testConfig] = buildspecs;
 
 	if (!testConfig) {
@@ -168,7 +169,7 @@ async function buildMultiEntry(targetPath) {
  * @param {string} targetPath
  */
 async function createScreenshots(targetPath) {
-	const buildspecs = getMatchingBuildspec(targetPath);
+	const buildspecs = await getMatchingBuildspec(targetPath);
 
 	const { url: cmdArgUrl, name, ...screenshotArgs } = cmdArgs;
 	let cmdArgBuild = false;
@@ -320,7 +321,7 @@ async function createScreenshots(targetPath) {
  * @param {string} targetPath
  * @return {Object[]}
  */
-function getMatchingBuildspec(targetPath) {
+async function getMatchingBuildspec(targetPath) {
 	let indexFiles = [];
 
 	targetPath = path.resolve(process.env.INIT_CWD, targetPath);
@@ -337,8 +338,8 @@ function getMatchingBuildspec(targetPath) {
 		);
 	}
 
-	return indexFiles
-		.map((entryFile) => {
+	const entries = await Promise.all(
+		indexFiles.map(async (entryFile) => {
 			const filter = createFilter([/\.(jsx?|tsx?|(le|sa|sc|c)ss)$/]);
 			if (entryFile.includes('.build')) {
 				return null;
@@ -346,11 +347,13 @@ function getMatchingBuildspec(targetPath) {
 			if (!filter(entryFile) && !lstatSync(entryFile).isDirectory()) {
 				return null;
 			}
-			const spec = buildspec(entryFile);
+			const spec = await buildspec(entryFile);
 			if (spec && new RegExp(`${spec.buildDir}(/|$)`).test(entryFile)) {
 				return null;
 			}
 			return spec;
 		})
-		.filter(Boolean);
+	);
+
+	return entries.filter(Boolean);
 }
