@@ -3,8 +3,10 @@
  * @property {*|null} current
  */
 
-import { hooks, isSame, onNextTick } from './utils/internal';
+import { config, hooks, isSame, onNextTick } from './utils/internal';
 import { renderVnode, patchVnodeDom } from './utils/render';
+
+config.h = true;
 
 /**
  * Initializes the reference object.
